% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{calculate_hex_size}
\alias{calculate_hex_size}
\title{Calculate hex size}
\usage{
calculate_hex_size(msoas_df, type = c("mean", "median"), efficiency = 40)
}
\arguments{
\item{msoas_df}{A data frame with each row being an MSOA. Must contain a
`shape_area` column.}

\item{type}{whether to return the cellsize figure relating to the mean MSOA
size or the median, or both. Defaults to both.}

\item{efficiency}{The efficiency level to aim for, when calculating
hex size}
}
\value{
a named integer vector of length 1 or 2.
}
\description{
For a df of MSOAs with a `shape_area` column, return hex sizes that
correspond to the mean and/or median actual MSOA area from the df.
The number returned is fit to be used as the `cell_size` parameter in
`create_hexgrid()` (which is then passed to the `cellsize` parameter in
`sf::st_make_grid`). This figure relates to the *distance across* (from top
to bottom with flat-topped hexagons) the grid hexagons generated.
}
