% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexgrid.R
\name{collect_hexes}
\alias{collect_hexes}
\title{Collect hexes}
\usage{
collect_hexes(data_inputs, msoa, compact = 1)
}
\arguments{
\item{data_inputs}{list created by previous run:
list(previous results, grid)}

\item{msoa}{a particular MSOA (with centroid point geography),
whose best hex location from the hex grid is to be calculated}

\item{compact}{the degree to which the choice of which hex to allocate to
the current MSOA hshould prefer to touch any existing hexes in the group
(eg a local authority group). 1 means that any hex touching 1 or more
existing hexes in the group will always be preferred to a hex that isn't
touching any previous hexes in the current group.}
}
\value{
a list: current results and the remaining grid
}
\description{
Collect hexes
}
