% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexgrid.R
\name{create_hexgrid}
\alias{create_hexgrid}
\title{Create hexgrid}
\usage{
create_hexgrid(area_bounds, msoas_list, cell_size, ...)
}
\arguments{
\item{area_bounds}{An sf object (eg `sfc` collection) that provides the
overall shape. With polygon boundaries.}

\item{msoas_list}{A regional list of MSOAs by LAD. With centroids.}

\item{cell_size}{The distance from top-bottom of each grid cell (metres).}

\item{...}{params to pass through to collect_hexes (currently just
`compact`).}
}
\value{
A list with four elements: 1: the original area bounds, 2: a base hexgrid cropped to those hexagons that intersect with `area_bounds`, 3: just the hexes from the grid that have been allocated to MSOAs, and 4: a list of MSOA hexagons grouped by LAD.
}
\description{
Combines grid creation function and hexgrid creation
}
