% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexgrid.R
\name{create_hexgrid_national}
\alias{create_hexgrid_national}
\title{Create a hexgrid for a group of regions or a whole nation}
\usage{
create_hexgrid_national(data_inputs, area_bounds, msoas_list, ...)
}
\arguments{
\item{data_inputs}{A list created by the previous run, or from `.init`.}

\item{area_bounds}{An sf object (eg `sfc` collection) that provides the
overall shape. With polygon boundaries.}

\item{msoas_list}{A regional list of MSOAs by LAD. With centroids.}

\item{...}{params to pass through to `collect_hexes` (currently just
`compact`).}
}
\value{

}
\description{
Each region calculations needs to take into account any hexes from a
national grid already claimed by previous regions. Uses a grid already
created by previous runs, or manually as an .`init`. Designed for use
within a `map` loop (hence returns `n` as `n + 1`, etc)
}
