% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorted_msoa_batches_by_lad.R
\name{create_sorted_msoa_batches}
\alias{create_sorted_msoa_batches}
\title{Create sorted MSOA batches}
\usage{
create_sorted_msoa_batches(msoa_centroids, lad_centroids, msoas_list_orig)
}
\arguments{
\item{msoa_centroids}{An \code{sf} tibble of MSOA centroids}

\item{lad_centroids}{A list of LAD centroids}

\item{msoas_list_orig}{An \code{sf} tibble of MSOAs (with boundary geometries),
sorted by density}
}
\description{
Uses helper functions \code{sort_by_init_proximity()} and
\code{sort_lads_list_by_density()} to create the grouped and sorted MSOA lists
ready for the hex calculation process.
}
