% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorted_msoa_batches_by_lad.R
\name{create_sorted_msoa_batches_by_lad}
\alias{create_sorted_msoa_batches_by_lad}
\title{Create sorted MSOA batches}
\usage{
create_sorted_msoa_batches_by_lad(lads_sf, msoas_sf, msoa_centroids)
}
\arguments{
\item{lads_sf}{sf tibble of LADs (with boundary geometries)}

\item{msoas_sf}{sf tibble of MSOAs (with boundary geometries)}

\item{msoa_centroids}{sf tibble of MSOA centroids}
}
\description{
Uses helper functions sort_by_init_proximity() and
sort_lads_list_by_density() to create the grouped and sorted MSOA lists
ready for the hex calculation process
}
