% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_current_region}
\alias{get_current_region}
\title{Just a chooser for one of the 9 English regions}
\usage{
get_current_region(regions_list, n)
}
\arguments{
\item{regions_list}{A tibble of 9 English regions}

\item{n}{numeric. Which region to choose (1-9, alphabetical order)}
}
\value{
The region name as a string
}
\description{
Just a chooser for one of the 9 English regions
}
