% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_msoa_centroids}
\alias{get_msoa_centroids}
\title{Get MSOA centroids}
\usage{
get_msoa_centroids(region)
}
\arguments{
\item{region}{name of region}
}
\value{
sf tibble of MSOAs in a region with centroid geometries
}
\description{
Get MSOA centroids
}
