% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_msoas}
\alias{get_msoas}
\title{Get MSOAs with boundaries}
\usage{
get_msoas(lads_list)
}
\arguments{
\item{lads_list}{A tibble of LADs including column of LAD names (strings)}
}
\value{
\code{sf} tibble of MSOAs with boundaries
}
\description{
Get MSOAs with boundaries
}
