% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_regional_lads}
\alias{get_regional_lads}
\title{Get regional LADs}
\usage{
get_regional_lads(region)
}
\arguments{
\item{region}{string. The name of region}
}
\value{
\code{sf} tibble of all LADs in a region, with boundaries
}
\description{
Get regional LADs
}
