% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap_creation_helpers.R
\name{hex_array}
\alias{hex_array}
\title{Hex array: the guts of create_hexgrid()}
\usage{
hex_array(data_inputs, msoas_list)
}
\arguments{
\item{data_inputs}{A list created by the previous run: previous results and
a grid}

\item{msoas_list}{A grouped list of MSOAs, sorted by LAD area or whatever}
}
\value{
A list: current results and a grid
}
\description{
Hex array: the guts of create_hexgrid()
}
