% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regional_grid.R
\name{make_regional_grid}
\alias{make_regional_grid}
\title{Make a regional grid}
\usage{
make_regional_grid(area, cell_size)
}
\arguments{
\item{area}{An \code{sf} object (eg sfc collection) that provides the overall
shape}

\item{cell_size}{numeric. The size of grid cells in metres.}
}
\value{
An \code{sf} grid object
}
\description{
Make a regional grid
}
