% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msoahexmap-package.R
\docType{package}
\name{msoahexmap-package}
\alias{msoahexmap}
\alias{msoahexmap-package}
\title{msoahexmap: Creates Hexagon Cartograms for English MSOAs}
\description{
The intention is to make a hexagon cartogram framework for MSOAs in England.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/msoahexmap}
  \item Report bugs at \url{https://github.com/francisbarton/msoahexmap/issues}
}

}
\author{
\strong{Maintainer}: Fran Barton \email{fbarton@alwaysdata.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

}
\keyword{internal}
