% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorted_msoa_batches_by_lad.R
\name{sort_by_init_proximity}
\alias{sort_by_init_proximity}
\title{Sort MSOAs by proximity to the closest ("initial") MSOA to the LAD centroid}
\usage{
sort_by_init_proximity(dtf)
}
\arguments{
\item{dtf}{sf tibble of MSOA centroids in a LAD}
}
\value{

}
\description{
Sort MSOAs by proximity to the closest ("initial") MSOA to the LAD centroid
}
