% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorted_msoa_batches_by_lad.R
\name{sort_lads_list_by_density}
\alias{sort_lads_list_by_density}
\title{Sort LADs list by density}
\usage{
sort_lads_list_by_density(msoas_list_by_proximity, msoas_list_orig)
}
\arguments{
\item{msoas_list_by_proximity}{A named list of MSOAs already split by LAD
name}

\item{msoas_list_orig}{An \code{sf} tibble of all MSOAs in area}
}
\value{
the original list, sorted
}
\description{
Use the "density" of MSOAs in each LAD as the way to order the hexgrid
creation/allocation process.
Density here means total LAD area divided by number of MSOAs in that LAD.
}
