% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap_creation_helpers.R
\name{collect_hexes}
\alias{collect_hexes}
\title{Collect hexes}
\usage{
collect_hexes(data_inputs, msoa)
}
\arguments{
\item{data_inputs}{A list created by the previous run: previous results and
a grid}

\item{msoa}{A particular MSOA whose hex location is to be calculated}
}
\value{
a list: current results and currently remaining grid
}
\description{
Collect hexes
}
