% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap_creation_helpers.R
\name{create_hexgrid}
\alias{create_hexgrid}
\title{Create hexgrid}
\usage{
create_hexgrid(area_bounds, msoas_list, cell_size)
}
\arguments{
\item{area_bounds}{An \code{sf} object (eg sfc collection) that provides the
overall shape}

\item{msoas_list}{A grouped list of MSOAs, sorted by LAD area or whatever}

\item{cell_size}{numeric. The size of grid cells in metres.}
}
\value{
A list with two elements: a base hexgrid and a hexgrid of MSOAs
}
\description{
Combines grid creation function and hexgrid creation
}
