% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hexgrid.R
\name{hex_array}
\alias{hex_array}
\title{Hex array: the guts of create_hexgrid()}
\usage{
hex_array(data_inputs, msoas_list, lads, ...)
}
\arguments{
\item{data_inputs}{A list created by previous run:
`list(previous results, grid)`.}

\item{msoas_list}{A list of MSOAs in a LAD, sorted by proximity to
LAD centroid or by shape_area.}

\item{lads}{total number of areas being processed in current batch}

\item{...}{Params to pass on to `collect_hexes` (currently just `compact`).}
}
\value{
A list: current results and a grid
}
\description{
Hex array: the guts of create_hexgrid()
}
