% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sorted_msoa_batches_by_lad.R
\name{sort_lads_list_by_density}
\alias{sort_lads_list_by_density}
\title{Sort LADs list by density}
\usage{
sort_lads_list_by_density(msoas_list_by_proximity, msoas_sf)
}
\arguments{
\item{msoas_list_by_proximity}{List of MSOAs already split by LAD name}

\item{msoas_sf}{sf tibble of all MSOAs in area}
}
\value{

}
\description{
Sort LADs list by density
}
